# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "md2",
    "description": "Converts a markdown document into a formatted binary document. The visual result is decided entirely by a theme, and the output format is selectable.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/convert": {
      "post": {
        "summary": "Convert a markdown document",
        "tags": [
          "conversion"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "markdown": {
                    "type": "string",
                    "minLength": 1
                  },
                  "theme": {
                    "type": "string",
                    "pattern": "^[a-z0-9][a-z0-9-]{1,31}$"
                  },
                  "format": {
                    "type": "string",
                    "pattern": "^[a-z0-9][a-z0-9-]{1,31}$"
                  },
                  "filename": {
                    "type": "string",
                    "maxLength": 200,
                    "pattern": "^[\\w\\-. ]+$"
                  },
                  "metadata": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string",
                        "maxLength": 500
                      },
                      "subtitle": {
                        "type": "string",
                        "maxLength": 500
                      },
                      "author": {
                        "maxItems": 20,
                        "type": "array",
                        "items": {
                          "type": "string",
                          "maxLength": 200
                        }
                      },
                      "date": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "subject": {
                        "type": "string",
                        "maxLength": 500
                      },
                      "keywords": {
                        "maxItems": 50,
                        "type": "array",
                        "items": {
                          "type": "string",
                          "maxLength": 100
                        }
                      },
                      "language": {
                        "type": "string",
                        "maxLength": 35
                      }
                    },
                    "additionalProperties": false
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "tableOfContents": {
                        "type": "boolean"
                      },
                      "titlePage": {
                        "type": "boolean"
                      },
                      "strict": {
                        "type": "boolean"
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "required": [
                  "markdown"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The converted document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "The converted document"
                }
              }
            }
          }
        }
      }
    },
    "/convert/{themeId}": {
      "post": {
        "summary": "Convert a raw markdown body with an explicit theme",
        "tags": [
          "conversion"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{1,31}$"
            },
            "in": "query",
            "name": "format",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 200,
              "pattern": "^[\\w\\-. ]+$"
            },
            "in": "query",
            "name": "filename",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "in": "path",
            "name": "themeId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The converted document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "The converted document"
                }
              }
            }
          }
        }
      }
    }
  },
...
}